+++ database/nostr-lmdb/src/store/mod.rs
@@ -27,7 +27,12 @@ pub struct Store {
 }
 
 impl Store {
-    pub(super) fn open<P>(path: P, map_size: usize) -> Result<Store, Error>
+    pub(super) fn open<P>(
+        path: P,
+        map_size: usize,
+        max_readers: u32,
+        additional_dbs: u32,
+    ) -> Result<Store, Error>
     where
         P: AsRef<Path>,
     {
@@ -36,7 +41,7 @@ impl Store {
         // Create the directory if it doesn't exist
         fs::create_dir_all(path)?;
 
-        let db: Lmdb = Lmdb::new(path, map_size)?;
+        let db: Lmdb = Lmdb::new(path, map_size, max_readers, additional_dbs)?;
         let ingester: Sender<IngesterItem> = Ingester::run(db.clone());
 
         Ok(Self { db, ingester })